+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Compatibility:
 * Support `#refine` for Module (#2021, @ssnickolay).
 * Implemented `rb_ident_hash_new`.
 * Improved the compatibility of `Symbol.all_symbols` (#2022, @chrisseaton).
+* Implemented `rb_enc_str_buf_cat`.
 
 Performance: